// src/models/CreateNamespacesBatchItemDto.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type CreateNamespacesBatchItemDto = {
    /**
     * Permission group unique identifier
     */
    code: string;
    /**
     *Permission group name
     */
    name?: string;
    /**
     * Permission group description information
     */
    description?: string;
};